http/tests/websocket/tests/hybi/close-on-* tests are not interoperable https://bugs.webkit.org/show_bug.cgi?id=215692
Reviewed by Darin Adler.

Before the patch, tests were not passing in Chrome, Firefox and NSURLSession WebSocket code path.
This was due to the fact the web socket script is using try/finally and the closure of the connection is not raising an exception
while it is for Chrome, Firefox and NSURLSession WebSocket code path.
Change the WebSocket server script to more reliably send the socket message at closing time of the connection.

* http/tests/websocket/tests/hybi/close-on-navigate-new-location-expected.txt:
* http/tests/websocket/tests/hybi/close-on-navigate-new-location.html:
* http/tests/websocket/tests/hybi/close-on-unload-and-force-gc-expected.txt:
* http/tests/websocket/tests/hybi/close-on-unload-and-force-gc.html:
* http/tests/websocket/tests/hybi/close-on-unload-expected.txt:
* http/tests/websocket/tests/hybi/close-on-unload-reference-in-parent-expected.txt:
* http/tests/websocket/tests/hybi/close-on-unload-reference-in-parent.html:
* http/tests/websocket/tests/hybi/close-on-unload.html:
* http/tests/websocket/tests/hybi/close-on-unload_wsh.py:
(web_socket_transfer_data):
* http/tests/websocket/tests/hybi/send-after-close-on-unload-expected.txt:
* http/tests/websocket/tests/hybi/send-after-close-on-unload.html:


Identifier: 228536@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266059 268f45cc-cd09-0410-ab3c-d52691b4dbfc