REGRESSION (r266028): platform/ios/ios/fast/coordinates/range-client-rects.html https://bugs.webkit.org/show_bug.cgi?id=215772
Reviewed by Anders Carlsson.

Source/WebCore:

* dom/Range.cpp:
(WebCore::Range::getClientRects const): Call updateLayout since it's not safe
to start working with the render tree without updating it first. Not required
to fix this bug, but an obvious omission. Ideally should make a test to show
this is needed.
(WebCore::Range::getBoundingClientRect const): Ditto.

* dom/SimpleRange.cpp:
(WebCore::order): Added.
(WebCore::documentOrder): Removed bogus special case. Test case proving this is
wrong is coming in the next documentOrder-related patch. Not needed to fix this
bug, but seems dangerous to leave this in the tree the way it was.
(WebCore::firstIntersectingNodeWithDeprecatedZeroOffsetStartQuirk): Added.
(WebCore::IntersectingNodeIterator::IntersectingNodeIterator): Added.

* dom/SimpleRange.h: Added intersectingNodesWithDeprecatedZeroOffsetStartQuirk.

* rendering/RenderObject.cpp:
(WebCore::borderAndTextRects): Use intersectingNodesWithDeprecatedZeroOffsetStartQuirk
because this needs to get the set of nodes in a way that works for our incorrectly
formed selection ranges and includes the first element even when the range technically
starts inside that element. We need to fix those, but that's a big project that affects
a lot of editing code. For now, restoring the old quirk is expedient.
(WebCore::RenderObject::collectSelectionRectsInternal): Ditto.

LayoutTests:

* platform/ios-wk2/TestExpectations: Removed failure expectation.

* platform/ios/ios/fast/coordinates/element-client-rects-expected.txt:
* platform/ios/ios/fast/coordinates/element-client-rects.html:
* platform/ios/ios/fast/coordinates/range-client-rects-expected.txt:
* platform/ios/ios/fast/coordinates/range-client-rects.html:
Changed tests so it's easier to read their results by putting rectangles and
rectangle lists into strings instead of comparing one value per line.

* platform/ios/ios/fast/coordinates/resources/helpers.js:
(rectString): Added.
(rectStrings): Added.
(setExpectedClientRectValues): Deleted.
(verifyClientRect): Deleted.


Identifier: 228598@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266123 268f45cc-cd09-0410-ab3c-d52691b4dbfc