[Mac] REGRESSION(r262322): Focusable elements are focused when exiting from video fullscreen mode. https://bugs.webkit.org/show_bug.cgi?id=215660 <rdar://problem/65263150>
Reviewed by Darin Adler.

Source/WebKit:

API Test: Fullscreen.Focus.

The default behavior of NSWindow is to change its -keyViewSelectionDirection to NSSelectingNext when activating
and changing the firstResponder. This causes WKWebView to change it's focus in response to becoming first responder.
The cheap fix is to ensure WKWebView is already the firstResponder when the window is activated. Then NSWindow won't
try to change the firstResponder in response to activation.

* UIProcess/mac/WKFullScreenWindowController.mm:
(-[WKFullScreenWindowController beganEnterFullScreenWithInitialFrame:finalFrame:]):

Tools:

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/mac/FullscreenFocus.mm: Added.
(-[FullscreenFocusUIDelegate _webViewDidEnterFullscreen:]):
(-[FullscreenFocusUIDelegate _webViewDidExitFullscreen:]):
(TestWebKitAPI::TEST):


Identifier: 228606@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266134 268f45cc-cd09-0410-ab3c-d52691b4dbfc