[css-grid] Set available column space before grid items prelayout https://bugs.webkit.org/show_bug.cgi?id=199648
Reviewed by Javier Fernandez.

LayoutTests/imported/w3c:

Import WPT test.

* web-platform-tests/css/css-grid/alignment/self-baseline/grid-self-baseline-008-expected.html: Added.
* web-platform-tests/css/css-grid/alignment/self-baseline/grid-self-baseline-008.html: Added.
* web-platform-tests/css/css-grid/alignment/self-baseline/w3c-import.log:

Source/WebCore:

Before running the grid track sizing algorithm, there is a prelayout
step which takes care of the baseline alignment of the items.
In order to do that, we estimate the size of the grid area, since we
have not calculated the final size yet.

The problem was that, when calculating this estimate, the available
column space had not been set yet. So if a column had a percentage track
sizing function, it would be treated as auto and the estimated size
would be wrong.

This patch sets the available column space so that we can resolve
percentage track sizing functions and get a more correct estimate.

Test: imported/w3c/web-platform-tests/css/css-grid/alignment/self-baseline/grid-self-baseline-008.html

* rendering/RenderGrid.cpp:
(WebCore::RenderGrid::layoutBlock):


Identifier: 228642@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266173 268f45cc-cd09-0410-ab3c-d52691b4dbfc