[JSC] setLength in Array#push could get very large length https://bugs.webkit.org/show_bug.cgi?id=215897 <rdar://problem/67859149>
Reviewed by Keith Miller.

JSTests:

* stress/array-push-more-than-max-size.js: Added.
(shouldBe):
(shouldThrow):

Source/JavaScriptCore:

Array#push can get length larger than UINT32_MAX. And in this case, we should throw a RangeError.
Before r266215, it was using putLength which throws an error. But it was replaced with setLength,
and JSC::setLength assumes that it never gets a length greater than UINT32_MAX by asserting. We
should fix it so that Array#push should thrown an error correctly.

* runtime/ArrayPrototype.cpp:
(JSC::setLength):

Identifier: 228709@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266257 268f45cc-cd09-0410-ab3c-d52691b4dbfc